+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This changelog tracks changes to the qoqo_iqm project starting at version 0.1.0.
 * Updated minimum supported Python version to 3.10.
 * Updated to pyo3 0.25.
 * Updated to qoqo 1.21 and qoqo_calculator 1.7.
-
+* Added the option to apply the virtual Z gate transformation before submitting the circuits.
 
 ## 0.10.2